requirements.txt
streamlit
PyPDF2
langchain
langchain-community
sentence-transformers
faiss-cpu
numpy